Apply unit tests
