+++ src/react/package.json
@@ -7,6 +7,7 @@
         "@craco/craco": "6.1.1",
         "@material-ui/core": "3.1.0",
         "@material-ui/icons": "3.0.1",
+        "@popperjs/core": "^2.11.8",
         "@reduxjs/toolkit": "^2.2.7",
         "@rollbar/react": "^0.11.2",
         "@turf/distance": "6.0.1",
